requirements.txt
matplotlib
numpy
pytest
pillow
scipy
